Added XML base stuffs
