+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@ package="com.example.banking_app">
             <category android:name="android.intent.category.LAUNCHER" />
         </intent-filter>
     </activity>
-    <activity android:name=".MainActivity"/>
+    <activity android:name=".activity.MainActivity"/>
     <activity android:name=".Signup"/>
 </application>
