+++ app/src/main/java/com/example/banking_app/Signup.java
@@ -11,10 +11,10 @@ import android.widget.TextView;
 import androidx.appcompat.app.AppCompatActivity;
 
 
-import java.io.IOException;
+import com.example.banking_app.activity.MainActivity;
+
 import java.sql.Connection;
 import java.sql.DriverManager;
-import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 public class Signup extends AppCompatActivity {
