+++ app/src/main/java/com/example/banking_app/fragments/OverviewFragment.kt
@@ -1,0 +1,16 @@
+package com.example.banking_app.fragments
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.banking_app.R
+
+class OverviewFragment : Fragment() {
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
+                              savedInstanceState: Bundle?): View? {
+        val view: View = inflater.inflate(R.layout.fragment_overview, container, false);
+        return view;
+    }
+}
