+++ app/src/main/res/layout/account_log.xml
@@ -1,0 +1,74 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+
+
+    <androidx.cardview.widget.CardView
+        android:layout_width="match_parent"
+        android:layout_height="123dp">
+
+        <TableLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+
+            <TableRow
+                android:layout_width="match_parent"
+                android:layout_height="match_parent" >
+
+                <de.hdodenhof.circleimageview.CircleImageView
+                    android:id="@+id/logRoundImage"
+                    android:layout_width="43dp"
+                    android:layout_height="match_parent"
+                    android:layout_marginStart="7dp"
+                    android:layout_marginEnd="2dp"
+                    android:padding="7dp" />
+                <TableLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent">
+                    <TableRow
+                        android:layout_width="match_parent"
+                        android:layout_height="match_parent" >
+
+                        <TextView
+                            android:layout_width="150dp"
+                            android:layout_height="46dp"
+                            android:text="TextView" />
+                        <TextView
+                            android:layout_width="150dp"
+                            android:layout_height="46dp"
+                            android:gravity="right"
+                            android:text="TextView" />
+                    </TableRow>
+
+                    <TableRow
+                        android:layout_width="match_parent"
+                        android:layout_height="match_parent"
+                        android:gravity="right"
+                        >
+
+                        <TextView
+                            android:layout_width="wrap_content"
+                            android:layout_height="42dp"
+                            android:paddingTop="10dp"
+                            android:gravity="right"
+                            android:text="TextView" />
+                    </TableRow>
+                </TableLayout>
+            </TableRow>
+
+
+            <TableRow
+                android:layout_width="match_parent"
+                android:layout_height="match_parent" >
+<!--here we will have the buttons/actions -->
+                <LinearLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:orientation="horizontal"></LinearLayout>
+            </TableRow>
+        </TableLayout>
+
+    </androidx.cardview.widget.CardView>
+</LinearLayout>
